remove now useless casts when invoking logger

+++ arangod/V8Server/v8-shape-conv.cpp
@@ -1042,7 +1042,7 @@ static v8::Handle<v8::Value> JsonShapeDataArray(v8::Isolate* isolate,
     }
 
     if (subshape == nullptr) {
-      LOG(WARN) << "cannot find shape #" << (unsigned int)sid;
+      LOG(WARN) << "cannot find shape #" << sid;
       continue;
     }
 
@@ -1050,7 +1050,7 @@ static v8::Handle<v8::Value> JsonShapeDataArray(v8::Isolate* isolate,
     char const* name = shaper->lookupAttributeId(aid);
 
     if (name == nullptr) {
-      LOG(WARN) << "cannot find attribute #" << (unsigned int)aid;
+      LOG(WARN) << "cannot find attribute #" << aid;
       continue;
     }
 
@@ -1075,7 +1075,7 @@ static v8::Handle<v8::Value> JsonShapeDataArray(v8::Isolate* isolate,
     }
 
     if (subshape == nullptr) {
-      LOG(WARN) << "cannot find shape #" << (unsigned int)sid;
+      LOG(WARN) << "cannot find shape #" << sid;
       continue;
     }
 
@@ -1083,7 +1083,7 @@ static v8::Handle<v8::Value> JsonShapeDataArray(v8::Isolate* isolate,
     char const* name = shaper->lookupAttributeId(aid);
 
     if (name == nullptr) {
-      LOG(WARN) << "cannot find attribute #" << (unsigned int)aid;
+      LOG(WARN) << "cannot find attribute #" << aid;
       continue;
     }
 
@@ -1152,7 +1152,7 @@ static v8::Handle<v8::Value> JsonShapeDataArray(v8::Isolate* isolate,
     }
 
     if (subshape == nullptr) {
-      LOG(WARN) << "cannot find shape #" << (unsigned int)sid;
+      LOG(WARN) << "cannot find shape #" << sid;
       continue;
     }
 
@@ -1160,7 +1160,7 @@ static v8::Handle<v8::Value> JsonShapeDataArray(v8::Isolate* isolate,
     char const* name = shaper->lookupAttributeId(aid);
 
     if (name == nullptr) {
-      LOG(WARN) << "cannot find attribute #" << (unsigned int)aid;
+      LOG(WARN) << "cannot find attribute #" << aid;
       continue;
     }
 
@@ -1185,7 +1185,7 @@ static v8::Handle<v8::Value> JsonShapeDataArray(v8::Isolate* isolate,
     }
 
     if (subshape == nullptr) {
-      LOG(WARN) << "cannot find shape #" << (unsigned int)sid;
+      LOG(WARN) << "cannot find shape #" << sid;
       continue;
     }
 
@@ -1193,7 +1193,7 @@ static v8::Handle<v8::Value> JsonShapeDataArray(v8::Isolate* isolate,
     char const* name = shaper->lookupAttributeId(aid);
 
     if (name == nullptr) {
-      LOG(WARN) << "cannot find attribute #" << (unsigned int)aid;
+      LOG(WARN) << "cannot find attribute #" << aid;
       continue;
     }
 
@@ -1252,7 +1252,7 @@ static v8::Handle<v8::Value> JsonShapeDataList(v8::Isolate* isolate,
     }
 
     if (subshape == nullptr) {
-      LOG(WARN) << "cannot find shape #" << (unsigned int)sid;
+      LOG(WARN) << "cannot find shape #" << sid;
       continue;
     }
 
@@ -1294,7 +1294,7 @@ static v8::Handle<v8::Value> JsonShapeDataHomogeneousList(
   subshape = shaper->lookupShapeId(sid);
 
   if (subshape == nullptr) {
-    LOG(WARN) << "cannot find shape #" << (unsigned int)sid;
+    LOG(WARN) << "cannot find shape #" << sid;
     return scope.Escape<v8::Value>(v8::Array::New(isolate));
   }
 
@@ -1343,7 +1343,7 @@ static v8::Handle<v8::Value> JsonShapeDataHomogeneousSizedList(
   subshape = shaper->lookupShapeId(sid);
 
   if (subshape == nullptr) {
-    LOG(WARN) << "cannot find shape #" << (unsigned int)sid;
+    LOG(WARN) << "cannot find shape #" << sid;
     return scope.Escape<v8::Value>(v8::Array::New(isolate));
   }
